# Conflicts:
#	types/canonical_types.go
#	types/commit.go
#	types/commit_test.go
#	types/evidence.go
#	types/proposal.go
#	types/validator.go
#	types/vote.go
#	types/vote_set.go

+++ types/evidence.go
@@ -188,7 +188,7 @@ func (dve *DuplicateVoteEvidence) String() string {
 
 // Height returns the height this evidence refers to.
 func (dve *DuplicateVoteEvidence) Height() uint64 {
-	return uint64(dve.VoteA.Height)
+	return dve.VoteA.Height
 }
 
 // Time return the time the evidence was created
